requirements.txt
yfinance
telebot
TA-Lib
python-dotenv